+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Bug Fixes
+
+- Fix multi-column sort by upgrade Avalonia to 0.9.3
+
 ## [1.4.0] - 2020-02-06
 
 ### New Features