+++ changelog.md
@@ -7,6 +7,10 @@
 - Allow textures to be loaded from a local directory
 - Allow remote textures to be cached locally for future use via config option `cacheRemoteTextures`
 
+### Enhancements
+
+- Add scrollbar to missable quests for users with high DPI settings
+
 ### Bug Fixes
 
 - Fix multi-column sort by upgrade Avalonia to 0.9.3