+++ changelog.md
@@ -16,6 +16,7 @@
 
 - Fix multi-column sort by upgrade Avalonia to 0.9.3
 - Fix crash when unable to reach remote texture path by providing a fallback texture
+- Fix not all summon cards being counted as obtained if their parent card was obtained
 
 ## [1.4.0] - 2020-02-06