+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [1.5.0] - 2020-04-07
 
 ### New Features
 
@@ -114,6 +114,7 @@
 ## [0.1.0-alpha] - 2017-01-26
 
 [unreleased]: https://github.com/rfvgyhn/gwent-tracker/compare/v1.4.0...HEAD
+[1.5.0]: https://github.com/rfvgyhn/gwent-tracker/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/rfvgyhn/gwent-tracker/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/rfvgyhn/gwent-tracker/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/rfvgyhn/gwent-tracker/compare/v1.1.0...v1.2.0